courier_service_API/requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==3.2.16
django-phonenumber-field==7.0.1
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.11.4
mccabe==0.7.0
phonenumbers==8.13.3
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
zipp==3.11.0
